+++ packages/Core/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added transaction api.  Not all adapters support transactions, each one will need to implement it's own transaction methods.
 - Added transaction support to built in adapters where possible.
 - Added throttle to count update to improve performance.
+- Added `serve-link` run script for `npm link` situations.
 
 ## [2.3.3] 5-20-2019
 - Fixed issue with rebuilding list type indexes